refactor [#310] 기존PerformanceDTO 이름 및 파일 위치 변경
- PerformanceDTO -> PerformancePreviewDTO
