+++ src/main/java/org/sopt/confeti/api/user/facade/dto/response/UserFavoritePerformancesDTO.java
@@ -1,12 +1,12 @@
 package org.sopt.confeti.api.user.facade.dto.response;
 
 import java.util.List;
-import org.sopt.confeti.domain.view.performance.PerformanceDTO;
+import org.sopt.confeti.domain.view.performance.application.dto.response.PerformancePreviewDTO;
 
 public record UserFavoritePerformancesDTO(
         List<UserFavoritePerformanceDTO> performances
 ) {
-    public static UserFavoritePerformancesDTO from(final List<PerformanceDTO> performances) {
+    public static UserFavoritePerformancesDTO from(final List<PerformancePreviewDTO> performances) {
         return new UserFavoritePerformancesDTO(
                 performances.stream()
                         .map(UserFavoritePerformanceDTO::from)
