+++ src/main/java/org/sopt/confeti/domain/view/performance/application/PerformanceService.java
@@ -6,6 +6,7 @@ import org.sopt.confeti.domain.view.performance.Performance;
 import org.sopt.confeti.domain.view.performance.PerformanceArtist;
 import org.sopt.confeti.domain.view.performance.PerformanceTicketDTO;
 import org.sopt.confeti.domain.view.performance.application.dto.response.PerformanceDTO;
+import org.sopt.confeti.domain.view.performance.application.dto.response.PerformancePreviewDTO;
 import org.sopt.confeti.domain.view.performance.infra.repository.PerformanceDTORepository;
 import org.sopt.confeti.domain.view.performance.infra.repository.PerformanceRepository;
 import org.sopt.confeti.global.common.constant.PerformanceType;
@@ -28,7 +29,7 @@ public class PerformanceService {
     private final PerformanceRepository performanceRepository;
 
     @Transactional(readOnly = true)
-    public List<PerformanceDTO> getFavoritePerformancesPreview(final long userId) {
+    public List<PerformancePreviewDTO> getFavoritePerformancesPreview(final long userId) {
         return performanceDTORepository.findFavoritePerformancesPreview(userId);
     }
