+++ src/main/java/org/sopt/confeti/domain/view/performance/application/dto/response/PerformancePreviewDTO.java
@@ -1,20 +1,20 @@
-package org.sopt.confeti.domain.view.performance;
+package org.sopt.confeti.domain.view.performance.application.dto.response;
 
 import org.sopt.confeti.global.common.constant.PerformanceType;
 
-public record PerformanceDTO(
+public record PerformancePreviewDTO(
         long typeId,
         PerformanceType type,
         String title,
         String posterPath
 ) {
-    public static PerformanceDTO of(
+    public static PerformancePreviewDTO of(
             final long typeId,
             final String type,
             final String title,
             final String posterPath
     ) {
-        return new PerformanceDTO(
+        return new PerformancePreviewDTO(
                 typeId,
                 PerformanceType.convert(type),
                 title,
